+++ src/components/Drink/Drink.css
@@ -1,5 +1,5 @@
 .drink {
-  width: 20%;
+  width: 25%;
   height: 100%;
 }
